Add useShouldWindowOptions
