+++ packages/components/src/Form/Inputs/Combobox/ComboboxContext.ts
@@ -66,7 +66,7 @@ export interface ComboboxContextProps<
   closeOnSelectPropRef?: MutableRefObject<boolean>
   readOnlyPropRef?: MutableRefObject<boolean>
   windowedOptionsPropRef?: MutableRefObject<boolean>
-  isScrollingRef?: MutableRefObject<boolean>
+  isAutoScrollingRef?: MutableRefObject<boolean>
   isVisible?: boolean
   openOnFocus?: boolean
   listScrollPosition?: number
