+++ packages/components/src/Form/Inputs/Select/SelectOptions.tsx
@@ -199,30 +199,8 @@ export function SelectOptions({
   formatCreateLabel,
   isMulti,
   noOptionsLabel = 'No options',
-  windowedOptions: propsWindowedOptions,
+  windowedOptions,
 }: SelectOptionsProps) {
-  const windowedOptions = useMemo(() => {
-    if (!options) return false
-    if (propsWindowedOptions === false) return false
-    // Without windowedOptions prop, default is to turn it on at 100 options
-    if (options.length < 100 && !propsWindowedOptions) return false
-    // But we can't use windowedOptions if there are groups
-    const groupedOptions = options.find(
-      (option) => (option as SelectOptionGroupProps).options !== undefined
-    )
-    if (groupedOptions) {
-      if (propsWindowedOptions) {
-        // If the windowedOptions prop is true but there are groups, give a warning
-        /* eslint-disable-next-line no-console */
-        console.warn(
-          'The `windowedOptions` prop does not support grouped options.'
-        )
-      }
-      return false
-    }
-    return true
-  }, [options, propsWindowedOptions])
-
   const {
     start,
     end,
