+++ packages/components/src/Form/Inputs/Select/utils/useWindowedOptions.tsx
@@ -28,12 +28,43 @@ import findIndex from 'lodash/findIndex'
 import React, { useContext, useEffect, useMemo, useRef } from 'react'
 import { getWindowedListBoundaries } from '../../../../utils/getWindowedListBoundaries'
 import { ComboboxContext, ComboboxMultiContext } from '../../Combobox'
-import { SelectOptionProps, SelectOptionObject } from '../SelectOptions'
+import {
+  SelectOptionGroupProps,
+  SelectOptionObject,
+  SelectOptionProps,
+} from '../SelectOptions'
 
 export const optionHeight = 28
 
+export function useShouldWindowOptions(
+  options?: SelectOptionProps[],
+  propsWindowedOptions?: boolean
+) {
+  return useMemo(() => {
+    if (!options) return false
+    if (propsWindowedOptions === false) return false
+    // Without windowedOptions prop, default is to turn it on at 100 options
+    if (options.length < 100 && !propsWindowedOptions) return false
+    // But we can't use windowedOptions if there are groups
+    const groupedOptions = options.find(
+      (option) => (option as SelectOptionGroupProps).options !== undefined
+    )
+    if (groupedOptions) {
+      if (propsWindowedOptions) {
+        // If the windowedOptions prop is true but there are groups, give a warning
+        /* eslint-disable-next-line no-console */
+        console.warn(
+          'The `windowedOptions` prop does not support grouped options.'
+        )
+      }
+      return false
+    }
+    return true
+  }, [options, propsWindowedOptions])
+}
+
 export function useWindowedOptions(
-  windowedOptions: boolean,
+  windowedOptions?: boolean,
   options?: SelectOptionProps[],
   isMulti?: boolean
 ) {
